requirements.txt
Flask
requests
Pillow
numpy
joblib
python-dotenv
gunicorn
